restructured cubit directories
