--- lib/blocs/multi_select_cubit/multi_select_state.dart
@@ -1,9 +1,0 @@
-part of 'multi_select_cubit.dart';
-
-class MultiSelectState {
-  final List<String> items;
-
-  const MultiSelectState({
-    required this.items,
-  });
-}
