+++ lib/ui/app_view.dart
@@ -4,14 +4,14 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:wakaranai/blocs/auth/authentication_cubit.dart';
 import 'package:wakaranai/generated/l10n.dart';
-import 'package:wakaranai/ui/anime_concrete_viewer/anime_concrete_viewer.dart';
-import 'package:wakaranai/ui/anime_iframe_player/anime_iframe_player.dart';
-import 'package:wakaranai/ui/anime_service_viewer/anime_service_viewer.dart';
+import 'package:wakaranai/ui/services/anime/anime_concrete_viewer/anime_concrete_viewer.dart';
+import 'package:wakaranai/ui/services/anime/anime_iframe_player/anime_iframe_player.dart';
+import 'package:wakaranai/ui/services/anime/anime_service_viewer/anime_service_viewer.dart';
 import 'package:wakaranai/ui/home/configs_page/my_repositores/my_repositores_page.dart';
 import 'package:wakaranai/ui/home/web_browser_page.dart';
-import 'package:wakaranai/ui/manga_service_viewer/concrete_viewer/chapter_viewer/chapter_viewer.dart';
-import 'package:wakaranai/ui/manga_service_viewer/concrete_viewer/manga_concrete_viewer.dart';
-import 'package:wakaranai/ui/manga_service_viewer/manga_service_viewer.dart';
+import 'package:wakaranai/ui/services/manga/manga_service_viewer/concrete_viewer/chapter_viewer/chapter_viewer.dart';
+import 'package:wakaranai/ui/services/manga/manga_service_viewer/concrete_viewer/manga_concrete_viewer.dart';
+import 'package:wakaranai/ui/services/manga/manga_service_viewer/manga_service_viewer.dart';
 import 'package:wakaranai/ui/routes.dart';
 import 'package:wakaranai/ui/splashscreen/splashscreen_view.dart';
 import 'package:wakaranai/utils/text_styles.dart';
