+++ lib/ui/home/configs_page/configs_group.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:wakaranai/data/models/remote_config/remote_category.dart';
 import 'package:wakaranai/data/models/remote_config/remote_config.dart';
-import 'package:wakaranai/ui/anime_service_viewer/anime_service_viewer.dart';
+import 'package:wakaranai/ui/services/anime/anime_service_viewer/anime_service_viewer.dart';
 import 'package:wakaranai/ui/home/configs_page/config_card.dart';
-import 'package:wakaranai/ui/manga_service_viewer/manga_service_viewer.dart';
+import 'package:wakaranai/ui/services/manga/manga_service_viewer/manga_service_viewer.dart';
 import 'package:wakaranai/utils/app_colors.dart';
 import 'package:wakaranai/utils/text_styles.dart';
 
@@ -11,8 +11,7 @@ import '../../routes.dart';
 
 class ConfigsGroup extends StatelessWidget {
   const ConfigsGroup(
-      {Key? key, required this.title, required this.remoteConfigs})
-      : super(key: key);
+      {super.key, required this.title, required this.remoteConfigs});
 
   final String title;
   final List<RemoteConfig> remoteConfigs;
