+++ lib/ui/home/settings/cubit/settings/settings_cubit.dart
@@ -1,11 +1,11 @@
 import 'package:bloc/bloc.dart';
 import 'package:meta/meta.dart';
-import 'package:wakaranai/blocs/remote_configs/remote_configs_cubit.dart';
 import 'package:wakaranai/env.dart';
 import 'package:wakaranai/data/models/configs_source_item/configs_source_item.dart';
 import 'package:wakaranai/data/models/configs_source_type/configs_source_type.dart';
 import 'package:wakaranai/services/settings_service/settings_service.dart';
-import 'package:wakaranai/ui/manga_service_viewer/concrete_viewer/chapter_viewer/chapter_view_mode.dart';
+import 'package:wakaranai/ui/home/configs_page/bloc/remote_configs/remote_configs_cubit.dart';
+import 'package:wakaranai/ui/services/manga/manga_service_viewer/concrete_viewer/chapter_viewer/chapter_view_mode.dart';
 
 part 'settings_state.dart';
