+++ lib/ui/services/anime/anime_service_viewer/anime_service_viewer.dart
@@ -9,11 +9,11 @@ import 'package:pull_to_refresh/pull_to_refresh.dart';
 import 'package:wakaranai/blocs/browser_interceptor/browser_interceptor_cubit.dart';
 import 'package:wakaranai/blocs/service_view/service_view_cubit.dart';
 import 'package:wakaranai/data/models/remote_config/remote_config.dart';
-import 'package:wakaranai/ui/anime_service_viewer/anime_service_viewer_body.dart';
 import 'package:wakaranai/ui/home/api_controller_wrapper.dart';
 import 'package:wakaranai/ui/home/service_view_cubit_wrapper.dart';
 import 'package:wakaranai/ui/home/web_browser_wrapper.dart';
 import 'package:wakaranai/ui/routes.dart';
+import 'package:wakaranai/ui/services/anime/anime_service_viewer/anime_service_viewer_body.dart';
 
 class AnimeServiceViewerData {
   final RemoteConfig? remoteConfig;
